Auto-add equals directive when on invokes/overrides mutations
When trying to apply e.g. an invokes directive, we will search among
all method declarations occuring in the template and pick one.
If that decl is neither replaced by a variable, nor has an equals
directive, we will add an equals directive, and use its metavariable
operand as the operand of the invokes directive.

By doing so, you can apply e.g. an invokes directive without having to
wait for a suitable equals directive to be added in an earlier mutation.