Implemented operators to ignore a node & to isolate an expression
add-directive-ignore: substitutes the subject for its first child,
during matching .. i.e. this node is skipped in the matching process

isolate-expr-in-method: substitutes a method body for just the selected
expression .. such that it matches with any method containing this
expression

.. these operators extend snippet-node-children|conceptually, a core
function of query generation! These extensions make it possible to
alter the tree structure during matching .. which consequently makes it
possible to express templates that don’t correspond to a valid Java AST
.. which is necessary to express e.g. “any method containing this
particular expression”