Can define template groups with pre-filled-in logic variables
- Can now define template groups where the values of logic vars is
provided in advance. This is useful in particular when generating
systematic code. (e.g. adding several similar fields + their uses)
- The values of these logic vars is specified as a Clojure expression
that must produce a list of strings, where each String is a Java
snippet.
Note that, in this Clojure expression, the value of previously defined
lvars can be used. (useful for cases such as: one lvar provides the
names of several fields, another var prefixes it with “get” to have the
getter methods’ names)
- Possible to define vector logic variables (the Clojure expr should
then return a list of lists of Strings) This is useful when certain
lvar values need to be paired (e.g. the type of a field + its name)
- The text parser now also supports the inlined wildcard and
metavariable syntax.